Working on processor API
