+++ trustgraph-base/trustgraph/base/__init__.py
@@ -1,8 +1,6 @@
 
-from . base_processor import BaseProcessor
+from . pubsub import PulsarClient
+from . async_processor import AsyncProcessor
 from . consumer import Consumer
 from . producer import Producer
-from . consumer_producer import ConsumerProducer
-from . publisher import Publisher
-from . subscriber import Subscriber
